refactor: common newMember component
