+++ apps/bilan-carbone/src/services/serverFunctions/user.command.ts
@@ -1,18 +1,6 @@
 import { Country, Role, SiteCAUnit } from '@abc-transitionbascarbone/db-common/enums'
 import z from 'zod'
 
-export const AddMemberCommandValidation = z.object({
-  email: z
-    .email()
-    .trim()
-    .transform((email) => email.toLowerCase()),
-  firstName: z.string().trim().min(1),
-  lastName: z.string().trim().min(1),
-  role: z.enum(Role),
-})
-
-export type AddMemberCommand = z.infer<typeof AddMemberCommandValidation>
-
 export const EditProfileCommandValidation = z.object({
   firstName: z.string().min(1).trim(),
   lastName: z.string().min(1).trim(),
